Return UnrecognizedFileFormat for files too small
Previously `pdb::PDB::open` would return a std::io::Error with ErrorKind
UnexpectedEof for files which were too small to be parsed into a PDB.
This EOF may indeed be a proper EOF if the file is small and therefore
we should return UnrecognizedFileFormat.

Adds a test to verify the behavior added here.